refactor: divide ssh connect and db connect
